+++ src/components/common/Description/Description.tsx
@@ -1,28 +1,28 @@
 import React, { useState, useEffect, useRef, useLayoutEffect } from 'react'
-import { patchApplicationNote, patchClusterNote } from '../ClusterNodes/clusterNodes.service'
+import { patchApplicationNote, patchClusterNote } from '../../ClusterNodes/clusterNodes.service'
 import ReactMde from 'react-mde'
 import 'react-mde/lib/styles/css/react-mde-all.css'
 import { showError, toastAccessDenied } from '@devtron-labs/devtron-fe-common-lib'
-import { MDEditorSelectedTabType } from '../ClusterNodes/types'
-import { ReactComponent as HeaderIcon } from '../../assets/icons/mdeditor/ic-header.svg'
-import { ReactComponent as BoldIcon } from '../../assets/icons/mdeditor/ic-bold.svg'
-import { ReactComponent as ItalicIcon } from '../../assets/icons/mdeditor/ic-italic.svg'
-import { ReactComponent as LinkIcon } from '../../assets/icons/mdeditor/ic-link.svg'
-import { ReactComponent as StrikethroughIcon } from '../../assets/icons/mdeditor/ic-strikethrough.svg'
-import { ReactComponent as CodeIcon } from '../../assets/icons/mdeditor/ic-code.svg'
-import { ReactComponent as QuoteIcon } from '../../assets/icons/mdeditor/ic-quote.svg'
-import { ReactComponent as ImageIcon } from '../../assets/icons/mdeditor/ic-image.svg'
-import { ReactComponent as OrderedListIcon } from '../../assets/icons/mdeditor/ic-ordered-list.svg'
-import { ReactComponent as UnorderedListIcon } from '../../assets/icons/mdeditor/ic-unordered-list.svg'
-import { ReactComponent as CheckedListIcon } from '../../assets/icons/mdeditor/ic-checked-list.svg'
-import { ReactComponent as DescriptionIcon } from '../../assets/icons/ic-note.svg'
-import { ReactComponent as DropDownIcon } from '../../assets/icons/ic-chevron-down.svg'
-import { ReactComponent as Edit } from '../../assets/icons/ic-pencil.svg'
+import { MDEditorSelectedTabType } from '../../ClusterNodes/types'
+import { ReactComponent as HeaderIcon } from '../../../assets/icons/mdeditor/ic-header.svg'
+import { ReactComponent as BoldIcon } from '../../../assets/icons/mdeditor/ic-bold.svg'
+import { ReactComponent as ItalicIcon } from '../../../assets/icons/mdeditor/ic-italic.svg'
+import { ReactComponent as LinkIcon } from '../../../assets/icons/mdeditor/ic-link.svg'
+import { ReactComponent as StrikethroughIcon } from '../../../assets/icons/mdeditor/ic-strikethrough.svg'
+import { ReactComponent as CodeIcon } from '../../../assets/icons/mdeditor/ic-code.svg'
+import { ReactComponent as QuoteIcon } from '../../../assets/icons/mdeditor/ic-quote.svg'
+import { ReactComponent as ImageIcon } from '../../../assets/icons/mdeditor/ic-image.svg'
+import { ReactComponent as OrderedListIcon } from '../../../assets/icons/mdeditor/ic-ordered-list.svg'
+import { ReactComponent as UnorderedListIcon } from '../../../assets/icons/mdeditor/ic-unordered-list.svg'
+import { ReactComponent as CheckedListIcon } from '../../../assets/icons/mdeditor/ic-checked-list.svg'
+import { ReactComponent as DescriptionIcon } from '../../../assets/icons/ic-note.svg'
+import { ReactComponent as DropDownIcon } from '../../../assets/icons/ic-chevron-down.svg'
+import { ReactComponent as Edit } from '../../../assets/icons/ic-pencil.svg'
 import Tippy from '@tippyjs/react'
-import { deepEqual } from '../common'
+import { deepEqual } from '..'
 import { toast } from 'react-toastify'
 import moment from 'moment'
-import { Moment12HourFormat } from '../../config'
+import { Moment12HourFormat } from '../../../config'
 import {
     DEFAULT_MARKDOWN_EDITOR_PREVIEW_MESSAGE,
     MARKDOWN_EDITOR_COMMANDS,
@@ -32,10 +32,10 @@ import {
     CLUSTER_DESCRIPTION_UPDATE_MSG,
     MARKDOWN_EDITOR_COMMAND_ICON_TIPPY_CONTENT,
     MARKDOWN_EDITOR_COMMAND_TITLE,
-} from '../ClusterNodes/constants'
+} from '../../ClusterNodes/constants'
 import './description.scss'
-import { MarkDown } from '../charts/discoverChartDetail/DiscoverChartDetails'
-import { AppMetaInfo } from '../app/types'
+import { MarkDown } from '../../charts/discoverChartDetail/DiscoverChartDetails'
+import { AppMetaInfo } from '../../app/types'
 
 export default function GenericDescription({
     isClusterTerminal,
