+++ src/components/common/Description/description.scss
